feat: Implement API for fetching room list
