+++ src/pages/multi/multi.tsx
@@ -1,10 +1,11 @@
 import { useState } from "react";
 import { CreateRoomModal } from "../../components/modal/room";
 import NavBar from "../../components/navbar/navbar";
-import { MainPageSearchBar, MainPageSearchBarWrap, MainPageTable, MainPageTableTbody, MainPageTableTbodyIcon, MainPageTableTbodyTd, MainPageTableTbodyTr, MainPageTableThead, MainPageTableTheadTh, MainPageTableTheadTr, MainPageTitleText, MultiPageTitle, MultiPageTitleIcon } from "./styled";
+import { MainPageSearchBar, MainPageSearchBarWrap, MainPageTable, MainPageTableThead, MainPageTableTheadTh, MainPageTableTheadTr, MainPageTitleText, MultiPageTitle, MultiPageTitleIcon } from "./styled";
 import { PrimaryButtonMedium } from "../../components/buttons/styled";
 import { useNavigate } from "react-router-dom";
 import { Background } from "../../components/background/styled";
+import RoomList from './roomlist/roomlist';
 
 export default function MultiPage(){
     const [isModalOpen, setIsModalOpen] = useState(false);
@@ -18,17 +19,6 @@ export default function MultiPage(){
       const handleDone = () => {
         setIsModalOpen(false);
       };
-      const handleRowClick = (roomId: string, roomName: string) => {
-        navigate(`/room/${roomId}`, { state: { roomName } });
-    }
-    
-
-    //임시 방 정보 하드코딩
-    const rooms = [
-        { id: "room1", name: "이기는 팀 우리 팀", locked: true, participants: "7/8" },
-        { id: "room2", name: "패배 팀", locked: false, participants: "5/8" },
-        { id: "room3", name: "우승 팀", locked: true, participants: "6/8" },
-    ];
 
     return(
         <>
@@ -50,19 +40,9 @@ export default function MultiPage(){
                             <MainPageTableTheadTh>인원</MainPageTableTheadTh>
                         </MainPageTableTheadTr>
                     </MainPageTableThead>
-                    <MainPageTableTbody>
-                        {rooms.map((room) => (
-                            <MainPageTableTbodyTr key={room.id} onClick={() => handleRowClick(room.id, room.name)}>
-                                <MainPageTableTbodyTd>{room.name}</MainPageTableTbodyTd>
-                                <MainPageTableTbodyTd>
-                                    {room.locked ? (
-                                        <MainPageTableTbodyIcon src="/icons/lock.svg" alt="Lock" />
-                                    ) : null}
-                                </MainPageTableTbodyTd>
-                                <MainPageTableTbodyTd>{room.participants}</MainPageTableTbodyTd>
-                            </MainPageTableTbodyTr>
-                        ))}
-                    </MainPageTableTbody>
+                    
+                    <RoomList/>
+
                 </MainPageTable>
             </Background>
             <CreateRoomModal
